normalize-yaml: Rename "format content" to "smart punctuation"
