+++ app/javascript/packages/normalize-yaml/cli.js
@@ -18,7 +18,7 @@ const options = {
   prettierConfig,
   exclude: /** @type {import('./index').Formatter[]} */ ([
     flags.includes('--disable-sort-keys') && 'sortKeys',
-    flags.includes('--disable-format-content') && 'formatContent',
+    flags.includes('--disable-smart-punctuation') && 'smartPunctuation',
   ].filter(Boolean)),
 };
