+++ app/javascript/packages/normalize-yaml/visitors/index.js
@@ -1,11 +1,11 @@
-import formatContent from './format-content.js';
+import smartPunctuation from './smart-punctuation.js';
 import sortKeys from './sort-keys.js';
 
 /** @typedef {import('yaml').visitor} Visitor */
 /** @typedef {import('../').Formatter} Formatter */
 
 /** @type {Record<Formatter, Visitor>} */
-const DEFAULT_VISITORS = { formatContent, sortKeys };
+const DEFAULT_VISITORS = { smartPunctuation, sortKeys };
 
 /**
  * @param {{ exclude?: Formatter[] }} exclude
